Implemented github api usage via okhttp and retrofit

+++ GithubUsers/app/src/main/java/com/rino/githubusers/screens/AndroidScreens.kt
@@ -7,5 +7,5 @@ import com.rino.githubusers.ui.users.UsersFragment
 
 class AndroidScreens : IScreens {
     override fun users() = FragmentScreen { UsersFragment.newInstance() }
-    override fun user(userId: Int): Screen = FragmentScreen { UserFragment.newInstance(userId) }
+    override fun user(userId: Long): Screen = FragmentScreen { UserFragment.newInstance(userId) }
 }
