+++ GithubUsers/app/src/main/java/com/rino/githubusers/screens/IScreens.kt
@@ -4,6 +4,6 @@ import com.github.terrakok.cicerone.Screen
 
 interface IScreens {
     fun users(): Screen
-    fun user(userId: Int): Screen
+    fun user(userId: Long): Screen
 }
